+++ src/NavDrawer/index.ts
@@ -1,5 +1,8 @@
-import NavDrawer, { DrawerItem as DrawerItemTemp } from "./nav_drawer";
+import NavDrawer from './nav_drawer';
+import {
+	DrawerItem as DrawerItemTemp,
+} from './nav_drawer';
+
 
 export default NavDrawer;
-
-export type DrawerItem = DrawerItemTemp
+export type DrawerItem = DrawerItemTemp;