+++ src/PageFrame/index.ts
@@ -1,6 +1,10 @@
-import PageFrame, { PageFrameDrawerOptions as PageFrameDrawerOptionsTemp, PageFrameNavbarOptions as PageFrameNavbarOptionsTemp } from "./page_frame";
+import PageFrame from './page_frame';
+import {
+	PageFrameDrawerOptions as PageFrameDrawerOptionsTemp,
+	PageFrameNavbarOptions as PageFrameNavbarOptionsTemp,
+} from './page_frame';
+
 
 export default PageFrame;
-
 export type PageFrameDrawerOptions = PageFrameDrawerOptionsTemp;
 export type PageFrameNavbarOptions = PageFrameNavbarOptionsTemp;